Working...

+++ clients/java/java-springboot/src/main/java/xcarpaccio/OrderCalculator.java
@@ -2,9 +2,10 @@ package xcarpaccio;
 
 public class OrderCalculator {
 
-    public static final int STANDARD_REDUCTION_1K_PLUS = 1 - 3 / 100;
-    public static final int STANDARD_REDUCTION_10K_PLUS = 1 - 10 / 100;
-    public static final int STANDARD_REDUCTION_50K_PLUS = 1 - 15 / 100;
+    public static final double STANDARD_REDUCTION_1K_PLUS = 1 - 3 / 100;
+    public static final double STANDARD_REDUCTION_10K_PLUS = 1 - 10 / 100;
+    public static final double STANDARD_REDUCTION_50K_PLUS = 1 - 15 / 100;
+
     public static final int STANDARD_REDUCTION_THRESHOLD_50K = 50000;
 
     public OrderCalculator() {
